+++ Sources/SwiftAgent/PromptContext.swift
@@ -2,6 +2,29 @@
 
 import Foundation
 
-public protocol PromptContext: Sendable, Equatable {}
+public struct PromptContext<Reference>: Sendable, Equatable where Reference: PromptContextReference {
+  public init(references: [Reference], linkPreviews: [PromptContextLinkPreview]) {
+    self.references = references
+    self.linkPreviews = linkPreviews
+  }
+  
+  public static var empty: Self {
+    Self(references: [], linkPreviews: [])
+  }
+  
+  public var references: [Reference] = []
+  public var linkPreviews: [PromptContextLinkPreview] = []
+}
 
-public struct EmptyPromptContext: PromptContext {}
+public protocol PromptContextReference: Sendable, Equatable {}
+public struct EmptyPromptContextReference: PromptContextReference {}
+
+public struct PromptContextLinkPreview: Sendable, Equatable {
+  public var title: String
+  public var description: String
+  
+  public init(url: URL) async throws {
+    self.title = ""
+    self.description = ""
+  }
+}